Move completion manager plugin to separate folder
