+++ spyder/plugins/completion/manager/__init__.py
