+++ spyder/plugins/completion/manager/decorators.py
