+++ spyder/plugins/completion/manager/plugin.py
