+++ spyder/plugins/projects/plugin.py
@@ -42,7 +42,7 @@ from spyder.plugins.projects.widgets.explorer import ProjectExplorerWidget
 from spyder.plugins.projects.widgets.projectdialog import ProjectDialog
 from spyder.plugins.completion.languageserver import (
     LSPRequestTypes, FileChangeType, WorkspaceUpdateKind)
-from spyder.plugins.completion.decorators import (
+from spyder.plugins.completion.manager.decorators import (
     request, handles, class_register)
